worker/requirements.txt
llama-index-core
llama-index-vector-stores-qdrant
llama-index-embeddings-ollama
watchdog
unstructured
python-pptx
docx2txt
pymupdf
pandas
